Install given target file but no status
In this scenario, the target has a file at the item path, but the tree
has no status for it. That means this call to Install.Analyze is the
first to visit a package item with the given path.

Install does two things:
- Record the file status in the tree.
- Report a conflict.

Later features will have Install analyze further and perhaps propose a
different Desired state.